Version 1.0.2
Revised a few things, fixed a bug

+++ app/src/main/java/com/felkertech/n/munch/Objects/HistItem.java
@@ -3,8 +3,8 @@ package com.felkertech.n.munch.Objects;
 public class HistItem {
     private int id;
     private long time;
-    private int amount;
-    public HistItem(int id, long time, int amount) {
+    private float amount;
+    public HistItem(int id, long time, float amount) {
         this.id = id;
         this.time = time;
         this.amount = amount;
